NXDRIVE-844: Add manual test
